fix: fixed and improved logic for second instances
